+++ test/extensions/filters/http/ext_authz/config_test.cc
@@ -31,7 +31,7 @@ TEST(HttpExtAuthzConfigTest, CorrectProtoGrpc) {
 
   testing::StrictMock<Server::Configuration::MockFactoryContext> context;
   EXPECT_CALL(context, localInfo()).Times(1);
-  EXPECT_CALL(context, clusterManager()).Times(2);
+  EXPECT_CALL(context, clusterManager()).Times(1);
   EXPECT_CALL(context, runtime()).Times(1);
   EXPECT_CALL(context, scope()).Times(2);
   EXPECT_CALL(context.cluster_manager_.async_client_manager_, factoryForGrpcService(_, _, _))
@@ -56,18 +56,24 @@ TEST(HttpExtAuthzConfigTest, CorrectProtoHttp) {
       allowed_headers: 
         patterns: 
         - exact: baz
-        - prefix: "x-"
+        - prefix: x-
+      headers_to_add:
+      - key: foo
+        value: bar
+      - key: bar
+        value: foo
+    
     authorization_response: 
       allowed_upstream_headers: 
         patterns: 
         - exact: baz
-        - prefix: "x-success"
+        - prefix: x-success
       allowed_client_headers: 
         patterns: 
         - exact: baz
-        - prefix: "x-fail"
+        - prefix: x-fail
 
-    path_prefix: "/extauth"
+    path_prefix: /extauth
     
   failure_mode_allow: true
   )EOF";